fix (db-utils) : have db-utils account for relative and absolute paths
Change default config to use import:/ instead of require:/ Shortstop when using
require:/ will not 'resolve' the protocols in included files, import:/ will.

Update dbUtils.createConfiguredConnections to account for absolute paths or
relative paths. When there is a shortstop protocol 'path:./path/file.json' it
will give the full path.

This now allows a koast server to not have it's own /server/schemas.js file,
and can use it's own default. If one is specified - will use that instead.

Allows developers to not need to use shortstop to specify paths to files.